trello-server/requirements.txt
# Core
Django==5.1.7
django-configurations==2.5.1
gunicorn==23.0.0
newrelic==10.3.1
setuptools==78.1.1

# For the persistence stores
psycopg2-binary==2.9.10
psycopg2==2.9.11
dj-database-url==2.3.0

# Model Tools
django-model-utils==5.0.0
django_unique_upload==0.2.1

# Rest apis
djangorestframework==3.15.2
django-filter==24.3

# Developer Tools
ipdb==0.13.13
ipython==8.30.0
mkdocs==1.6.1
flake8==7.1.1

# Testing
mock==5.1.0
factory-boy==3.3.1
pytest-django==4.9.0
coverage==7.6.9

# Static and Media Storage
django-storages==1.14.4
boto3==1.35.76

# Organization
django-organizations

# Redis
django-redis

# Cors
django-cors-headers

# JWT
djangorestframework-simplejwt